modularization of the tab system
